break get_oodle into separate package
