+++ Cargo.toml
@@ -14,6 +14,7 @@ aes = "0.8.2"
 base64 = "0.21.0"
 strum = { version = "0.24", features = ["derive"] }
 sha1 = "0.10"
+hex = "0.4"
 
 # generated by 'cargo dist init'
 [profile.dist]
