+++ get_oodle/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "get_oodle"
+repository.workspace = true
+authors.workspace = true
+license.workspace = true
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+repak = { path = "../repak" }
+sha1 = { workspace = true }
+ureq = "2.6"
+once_cell = "1.17"
+hex-literal = "0.4"
+libloading = "0.7"
+hex = { workspace = true }
