+++ repak_cli/Cargo.toml
@@ -12,20 +12,16 @@ path = "src/main.rs"
 
 [target.'cfg(windows)'.dependencies]
 repak = { path = "../repak", features = ["oodle"] }
-sha1 = { workspace = true }
-ureq = "2.6"
-once_cell = "1.17"
-hex-literal = "0.4"
-libloading = "0.7"
 
 [target.'cfg(not(windows))'.dependencies]
 repak = { path = "../repak" }
 
 [dependencies]
+get_oodle = { path = "../get_oodle" }
 aes = { workspace = true }
 base64 = { workspace = true }
 clap = { version = "4.1.4", features = ["derive"] }
-hex = "0.4.3"
+hex = { workspace = true }
 indicatif = { version = "0.17.3", features = ["rayon"] }
 path-clean = "0.1.0"
 path-slash = "0.2.1"
